python/requirements.txt
pandas>=2.0
matplotlib>=3.7
# optional
numpy>=1.24
# optional for DB export
psycopg2-binary>=2.9
